Update to FlipOutTests.js - extensive testing of many different things
I have updated this file to make tests and assertions of just about every circumstance I can think of.  The 'truffle-assertions' does not seem to be working correctly, so there a few tests with commented out lines where I have made a '.fails' assertion that hasn't worked properly (doesn't work with a '.fails' or a '.passes' assertion) so I have just made the rest of the test with standard assertions.  Everything passes as I expect it to but I obviously have not been able to check for any failures properly.